requirements.txt
django
djangorestframework
#django-rest-auth
django-filter
pygments
markdown
